Rewrite masters and procedures query to querydsl

+++ barbershop/src/main/java/com/example/barbershop/dtos/ProcedureDto.java
@@ -4,5 +4,6 @@ public interface ProcedureDto {
     Integer getProcedureId();
     String getName();
     Integer getDuration();
-    Double getPrice();
+    Integer getPrice();
 }
+
